phidgets_ik: Fix warnings

+++ phidgets_ik/src/ik_ros_i.cpp
@@ -69,7 +69,7 @@ void IKRosI::sensorHandler(int index, int sensorValue)
   CPhidgetInterfaceKit_getSensorRawValue(ik_handle_, index, &rawval);
   std_msgs::Float32 msg;
   msg.data = VREF*float(rawval)/4095.0f;
-  if ((sensor_pubs_.size() > index) && (sensor_pubs_[index])) {
+  if ((static_cast<int>(sensor_pubs_.size()) > index) && (sensor_pubs_[index])) {
     sensor_pubs_[index].publish(msg);
   }
 }
@@ -80,7 +80,7 @@ void IKRosI::inputHandler(int index, int inputValue)
   IK::inputHandler(index, inputValue);
   std_msgs::Bool msg;
   msg.data = inputValue != 0;
-  if ((in_pubs_.size() > index) && (in_pubs_[index])) {
+  if ((static_cast<int>(in_pubs_.size()) > index) && (in_pubs_[index])) {
     in_pubs_[index].publish(msg);
   }
 }
